+++ include/rhonabwy.h
@@ -299,7 +299,7 @@ void r_jwt_free(jwt_t * jwt);
  * @param alg: the algorithm to convert
  * @return the converted jwa_alg, R_JWA_ALG_NONE if alg is unknown
  */
-jwa_alg str_to_jwa_alg(const char * alg);
+jwa_alg r_str_to_jwa_alg(const char * alg);
 
 /**
  * Get the algorithm string
@@ -307,14 +307,14 @@ jwa_alg str_to_jwa_alg(const char * alg);
  * @param alg: the algorithm to convert
  * @return its string name
  */
-const char * jwa_alg_to_str(jwa_alg alg);
+const char * r_jwa_alg_to_str(jwa_alg alg);
 
 /**
  * Get the jwa_enc corresponding to the string algorithm specified
  * @param enc: the algorithm to convert
  * @return the converted jwa_enc, R_JWA_ENC_NONE if enc is unknown
  */
-jwa_enc str_to_jwa_enc(const char * enc);
+jwa_enc r_str_to_jwa_enc(const char * enc);
 
 /**
  * Get the encryption string
@@ -322,7 +322,7 @@ jwa_enc str_to_jwa_enc(const char * enc);
  * @param enc: the encryption to convert
  * @return its string name
  */
-const char * jwa_enc_to_str(jwa_enc enc);
+const char * r_jwa_enc_to_str(jwa_enc enc);
 
 /**
  * @}